Add the choice to have see-through labels in plots.py
Good day. 
I have been working on a script to make the labels see-through and just overall better looking in my opinion. To see the comparison please refer to Propose see-through labels. #4335
This is a boolean flag that is set to False by default so it won't change anything unless someone wants it then they can activate it by setting that flag to True. In my opinion, the see-through labels make it look much more elegant and would be a nice feature to have.