Update detect.py
As of now, the detect.py is loading the model at every single run.
This step is very costly. It can take more than 10 sec on devices with limited computing power (Nvidia Jetson Nanos, raspberry pi, ...)
As a consequence, using yolov5 on these devices is not realistic.
This PR proposes a method to improve the performance of the detect.py on low-power devices.
It does that by splitting the model loading, dataset loading, and inference into 3 different parts.
That way, the dataset, and inference can be completed without doing the expensive model loading step again.
It was done while not impacting how the current repository works.